move TestScenario to its own module
